+++ Sources/CoreCollector/adapters/PlayerContext.swift
@@ -3,4 +3,5 @@ import CoreMedia
 public protocol PlayerContext {
     var position: CMTime? { get }
     var isLive: Bool? { get }
+    var isPlaying: Bool { get }
 }