+++ Sources/CoreCollector/stateMachines/StateMachineFactory.swift
@@ -1,0 +1,5 @@
+public class StateMachineFactory {
+    public static func create(playerContext: PlayerContext) -> StateMachine {
+        DefaultStateMachine(playerContext: playerContext)
+    }
+}