+++ Sources/CoreCollector/stateMachines/util/HeartbeatListener.swift
@@ -1,0 +1,7 @@
+internal protocol PlayingHeartbeatListener: AnyObject {
+    func onPlayingHeartbeat() -> Bool
+}
+
+internal protocol RebufferHeartbeatListener: AnyObject {
+    func onRebufferHeartbeat()
+}